pei-data-engineering-project/requirements.txt
pyspark==3.4.0
pandas==2.0.0
openpyxl==3.1.2
pytest==7.4.0
pytest-cov==4.1.0
pyyaml==6.0
